refactor!: moved server methods into separate files and reorganised file structure
